[taint] Do not add taint attribute when the value is nil
Summary:
Before we were removing the Tainted attribute when the value was nil, but we also need to not add it if the value is nil. Otherwise the attribute disappears after some method call and debugging becomes difficult.

The log message should also help explain why some source is not acknowledged because the value is null.

Reviewed By: ngorogiannis

Differential Revision: D44914047

fbshipit-source-id: 1c1a2e486d24bf89467d8dff9099a2ddbabf6049