[textual] Allow only a restricted version of ... syntax in declarations
Summary:
Previously, it was possible to mix formals' types and ellipsis in textual function declarations to
supply **partial declarations**, e.g.
```
declare foo(int, bool, ...) : void
```

Although, this allowed for some extra flexibility, in practice, the mix of types and ellipsis wasn't
used. The only variation we actually use is `foo(...)` in Hack. The fledgling Python frontend also
doesn't seem to have a good use-case for partial decls.

With that said, this diff proposes reducing the state space by allowing only a limited form of
ellipsis syntax. This will also simplify the work on overloads support for Hack.

Reviewed By: vsiles

Differential Revision: D45127453

fbshipit-source-id: 7576e468a7e976bb10349d8cac3dacc104070945