[hack] Use arity in Hack procnames for overload resolution
Summary:
Overloads in Hack are a result of translation of functions with default arguments. Each such
overload can be uniquely identified by the function name and its arity.

Unfortunately, it's also possible that the function will remain in the tenv without an
arity (unknown decls). This in turn may require us to conjure a procname from the call-site using
the shape of the call in order to properly resolve cross-file calls to overloads.

Reviewed By: jvillard

Differential Revision: D45182413

fbshipit-source-id: 1fc83d633b3853e54ef72ec5f4bc4269982f7f06