[refactor] Lineage: make the Domain more abstract
Summary:
This commit refactors Lineage mostly by making `Domain` abstract (hide the type and provide the
updating functions directly in the module). This is to make sure that no code from outside the module
can break the invariants, especially that [LastWrites] should always be updated when adding flow into
a variable.

Reviewed By: rgrig

Differential Revision: D44139490

fbshipit-source-id: 7887de17f90673413ac7b8b1ad9a7b4a78f71f9a