[refactor] Lineage: instantiate all phantom parameters of VariableIndex.t
Summary:
This is the first step of a refactoring of Lineage "Variable Indices" best understandable as a series
of small atomic diffs.

This one removes all polymorphic occurrences of the `'phantom VariableIndex.t` type, which were
actually only instantiated to transient indices anyway, with the exception of `pp` which is only used
on terminal indices. We also define a `Transient` module to make some notations easier.

Reviewed By: rgrig

Differential Revision: D44351662

fbshipit-source-id: c27b13de5ea6ce02a16eb5254bbf370913243237