MinGW support, take 2 (#1755)
Summary:
This is a new iteration on https://github.com/facebook/infer/issues/1394. The code in that one was so old, and has changed so much, that it did not seem worth it to push on top of it. Some comments may still apply, and I apologize in advance for the (partial) double review.

This MR is not complete, in the sense that the code that uses Core_kernel instead of Core is not here. We are too far apart from you (in terms of OCaml and Core versions) for our code to be helpful. But the "meat" of the functionality is in this MR.

Pull Request resolved: https://github.com/facebook/infer/pull/1755

Reviewed By: dulmarod

Differential Revision: D45085006

Pulled By: jvillard

fbshipit-source-id: 7df3b4bb26849c37d9ba02bcc85ea705186e521e