[win] restore progress bar in --no-unix-fork + wibbles
Summary:
The auto-detection for the multiline progress bar was broken with
`--no-unix-fork` in the children processes because `stdout` is not a tty in
that case. As a consequence, by default the multiline progress bar was selected
by the orchestrator process but it didn't get the correct updates from its
children.

The main fix here is to make sure the auto-detection of the progress bar in the
originator process is transmitted down to children processes.

Some other cosmetic changes:
- the closure passed from originator to children can be broader and only take
  as argument what is absolutely needed, i.e. known only once the child process
  has been spawned (arguably nothing is needed since the arguments are the in and
  out file descriptors which are constants anyway since they are stdin and
  stdout, i.e. fds 0 and 1, but meh).
- pass all args to children via `INFER_ARGS` for consistency / slightly simpler
  code
- do not `log_environment_info ()` from children to reduce log spam
- put logic for running as a child in the same match block as other behaviours

Reviewed By: skcho

Differential Revision: D45314341

fbshipit-source-id: 78e4e1509d4a7259a8f92de6960bc5c25c5e34ec