[topl] Keep vars mentioned by path condition
Summary:
Consider a variable X that takes values v1, v2, and then v3. Topl state
may contain a constraint v2=42. Pulse state may be {pre: v1<0; post:
ret=v3; path_condition: v2=v1+1 & v3=v2+1}; that is; v2 is not mentioned
in either pre or post, yet its value is constrained (to not be 42) by
the path-condition.

In general, we need to keep Topl constraints that mention anything that
could be in any way constrained by the Pulse state: pre, post or
path-condition. Previously, the abstract values mentioned in the
path-condition where ignored.

Reviewed By: jvillard

Differential Revision: D45282247

fbshipit-source-id: a654d1cb551a39aa6a97e5cd8c55b6ea45170c97