[infer/python] refactor jump payload
Summary:
In preparation for the loop support, I refactor the `JUMP` logic so that
each command can provide the label information right away. I mostly moved code from `until_terminator` to the relevant opcode `run` functions.

The logic itself is not changed yet

Every loop in Python is an iterator, which is basically a TwoWay jump.
The difference is that in the "next" case, the element fetched from the
iterator must be pushed on the stack before running the body of the
loop. With this refactor, it will be easier to attach some instructions
to the label_info, so we can insert them between the pruning and the
actual node.

Reviewed By: jvillard

Differential Revision: D45269127

fbshipit-source-id: 0bc885ea99645801f908c97cc742847c83ab5e66