[taint] Refactor the taint modules
Summary:
The module PulseTaintOperations contained too many things, including data and functions over it, and I thought it would be good to separate things a bit. So far is just moving things around, but I'm planning to also add pp functions to the TaintConfig elements and print somewhere in debug mode what config we are looking at, which I think it's useful sometimes when debugging.

I called `TaintItem` the unit of taint that we carry around in attributes and reports, the name `Taint` seems too generic for that.

Then `Unit` is the unit of config that is a source or sink or sanitizer. Later we may want to separate out those given that there are items that can only be in a source, etc, and then maybe the parsing and error messages can also be improved.

Reviewed By: skcho

Differential Revision: D45527578

fbshipit-source-id: dee6fa2fc6039e9f4ff7034a1aa959d38e44ba40