Rename SimpleLineage into Lineage, SimpleShape into LineageShape
Summary:
Lineage was named SimpleLineage because it did not have context, path, nor field sensitivity. The
latter is not true anymore, and the former are not absent by "design" as we have some ideas to
integrate them.

Doing the renaming will also slightly simplify the CLI invocation of Lineage, which often involves
a set of options that are becoming long to type.

Regarding compatibility, the old option flags remain functional, albeit deprecated. The results are
still being exported in `infer-out/simple-lineage`. The json export filenames are now
`lineage-blah.json` instead of `simple-lineage-blah.json`, but we are not aware of any system relying
on the old file name format.

Reviewed By: ddino

Differential Revision: D45615018

fbshipit-source-id: 469ba37080461adc8a6fa40c5c9cbf4a4ecba812