[hack] adding static method inheritance
Summary:
Hack static methods can be overriden. We add a hiearchy walkup to find
implementation.

Before this diff, any declared but not implemented function was added to the list of
methods of a class in Tenv. We change this behavior and only record implemented methods.
If a methods can not be found, we expect it to be a builtin or a member of HackMixed.

Reviewed By: artempyanykh

Differential Revision: D45112415

fbshipit-source-id: 1c2a9e20652b1babf9a5c71ab0d3a66642200834