[pulse][specialization] first version of devirtualization by summary specialization
Summary:
We add a second type of summary specialization here. When a callee contains
a virtual call that can not be devirtualized by lack of information on the dynamic type of
the receiver, the callee can ask the caller to give this informatino trough a specialization.

We only support simplified cases frow now:
1) if the callee does not have this informatino, we give up, while we could ask the caller
of the caller.
2) if the receiver of the virtual call is not a parameter of the callee, we give up, while
we could express specialization needs in term of paths in the heap

Algorithm:
 - in the callee, if we encounter a virtual call where the dynamic type of the receiver is unknown, we search a formal (pvar) that contain this receiveir and put in our `need_dynamic_type_specialization` set (if such a pvar exist).
 - this set is recorded in the main summary of the callee (the other summaries are specialized versions)
 - when the caller calls the callee, it first looks into the callee's set. If it is not empty, he should gives information about the dynamic types of the corresponding callee arguments.

Reviewed By: jvillard

Differential Revision: D45314254

fbshipit-source-id: 91440c62d49bbf6e5ce5e5c3de7c9d7d85b28d30