[hack][tainting] Level5 unlocked
Summary:
`Level5.hack` is a Hack program proposed by Artem last year. It exercises various Hack OO features like inheritance, dynamic dispatch, static late binding.

Using the pulse `--pulse-force_continue`, and all companion diffs in this stack, we are now able to handle it properly.

This option is used because of our current handling of Hack abstract methods. Currently HackC generates a dummy method for them but these methods generate empty summary, and dynamic type specialization is weak in this situation. We need a bit more experiment to decide if we are happy with using this option always, or if we want to improve the handling of abstract methods.

Reviewed By: nbenton

Differential Revision: D45479546

fbshipit-source-id: abf5c98ddfee99a5c6533d333894270157b7870d