[hack] Clean up models a bit
Summary:
Having `ret null` as a model is not the best, especially for predicate
functions (e.g. `is_type_struct_c`) because it confuses the analysis with false nullptr dereferences
which is especially problematic when `pulse-force-continue` is not set.

This diff uses `nondet` for predicate functions. We could delete models but then builtins would
accumulate in `skipped_calls`.

I also removed a few `ret null` models where it wasn't easy to write something reasonabe.

NOTE: that there are 2 `ret null` models that we can't delete because otherwise we get a bunch of
FPs in tests. These need to be looked at.

Reviewed By: davidpichardie

Differential Revision: D45852568

fbshipit-source-id: 3671bd024e268ce4a474978774e66435b7bd34f7