[taint] Introduce the notion of field matchers, process parsing
Summary:
We want to be able to specify the action "stored a source in a given field/instance variable" as a sink in our taint analysis. So far everything in the code assumed that the sinks are procedures, so this requires a refactoring.

In this diff, as a first step towards our goal, we are able to parse field matchers that can be sinks.

In following diffs we will make use of them to find taint errors.

Reviewed By: geralt-encore

Differential Revision: D45866437

fbshipit-source-id: 0b1e7bc3b1964e7e07d4460fbad7bf9315370358