[debug] Print procnames in debug html verbosely
Summary:
Having verbose names simplifies navigating the HTML. In particular, in Hack there may be a bunch of
functions with the same name but different arity (due to default args); but we print arity only in
verbose mode.

I also tried Hack.pp to print arity in non-verbose mode too, but then CFG and test reportes become a
bit too cluttered with the arity info.

Reviewed By: jvillard

Differential Revision: D45984993

fbshipit-source-id: 0ee3bf283cd5d65534549bbbfd7abff81d1743d7