[debug] Add a control to make node instruction listing stick to the top
Summary:
Often when debugging real programs I need to cross-reference what those
`n$13` and `n$24` correspond to (so going up and down html).

Sticky instruction listing should simplify navigating and
cross-referencing.

Two button control whether the listing will stick to top and its visibility. For instance, it's OK to make the listing stick to the top and toggle its visibility when needed so that the listing doesn't obscure the output all the time.

Reviewed By: jvillard

Differential Revision: D46019382

fbshipit-source-id: c5cadc2b1a2d1c0ef8bcda232996acfddc8cc701