[infer/python][1/2] update python integration to allow capture of multiple files
Summary:
Disclaimer: this diff is heavily inspired by `integration/Hack.ml`.

I start to make infer capture and turn multiple python source files in sil.
This diff is the first of two which does the plumbing in `integration`. However some parts are missing (see error in Test plan) and I'll update the Sil part in the next diff.

Reviewed By: artempyanykh

Differential Revision: D46215595

fbshipit-source-id: 1e4c12e979af53768754b6dff97eb96488ddf71c