[scheduler] new "replay" scheduler
Summary:
The new scheduler is available with `--scheduler replay` and will build a graph
of work to do that follows the analysis dependencies computed with a previous
analysis. The goal is to obtain a second analysis that will mimic the results of the
initial one to reduce non-determinism due to cycles of mutual recursion in the
call graph. However that goal is not reached yet and there is still variability in
consecutive runs of the analysis even with `--scheduler replay` for now.

This assumes the previous analysis covered the exact same code, pretty much.
In particular, if there was no analysis previously it will not schedule any
work, so there's more to do to make it a bit more sensible. Eventually it can
interact with incremental analysis too to make infer differential runs more
deterministic.

Reviewed By: bennostein

Differential Revision: D46227133

fbshipit-source-id: 389fded9d99e1d1e1ab994893c6023a37b48899c