[infer/python] rewrite method parsing in a more systematic way
Summary:
In the previous diff, we support a minimal set of features for class declaration, but it was not enough to capture it's full generality. In this diff, I replace the old code with a small state machine which do a best effort parsing of the class declaration code.

We also add the parsing of type annotation for class members.

Note: we do not yet support method decorators (like `staticmethod`) as it we don't support closures and it will require it.
Also we turn types into strings too early. A future diff will improve this so we can correctly deal with generics.

Reviewed By: jvillard

Differential Revision: D46112352

fbshipit-source-id: eb8bf41fdba5982e99a489fba6010f56c4b5945c