[pulse] massage module types
Summary:
This changes the way the types of the main maps used in the abstract domain
(stack, heap, attributes) to introduce a distinction between the datatype
coming into the maps and the datatype coming out of the map. This is to make a
later commit easier and does not actually change the final OCaml types.

The main idea will be that values (`AbstractValue.t`) in those maps are always
normalised when they are *keys* but not necessarily when they are in *value*
position. The abductive domain will then normalise values on the fly while
maintaining the invariant that keys are normalised. See later commit.

Reviewed By: skcho

Differential Revision: D45917439

fbshipit-source-id: 9e8a8665fa1e8b618b7a7834711c75e9bfc49e45