[ondemand] rename file-analysis functions
Summary:
Makes it clearer how each function in the ondemand API is intended to be used.

Also separate the analysis call graph computation from the task scheduling one
in InferAnalyze just because it's a little more pleasant to read (and minimises
the amount of work we do before computing the actual scheduler to use).

Reviewed By: skcho

Differential Revision: D46479505

fbshipit-source-id: a3addebd6b08952144ffd1eba66fb7c09fd4c95b