[ondemand] replay recursion cycle breaks
Summary:
When ondemand detects a mutual recursion cycle, it returns an empty summary and
records that this happened in the caller procedure's dependencies. We can use
this recorded information to replay the same decisions in the replay scheduler.
This makes the analysis results deterministic! Though for some yet-unknowned
reason the analysis times can still vary.

Reviewed By: skcho

Differential Revision: D46479503

fbshipit-source-id: bf1f1c2e9c3568807a6c735a2d158a6f2025eefb