[infer/hack] Adding Hack class information to IR.Struct
Summary:
To correctly perform method resolution, we are going to need to know
which class is a trait and which class is an actual class. To this goal,
I extend the `IR.Struct` type with a Hack specific payload. For now we
only track trait vs class, but we'll probably extend it further in the
future

Reviewed By: ngorogiannis

Differential Revision: D46480484

fbshipit-source-id: f7272ea8eabef93096850d8a2ea052c2c20e9b10