[infer/hack][experimental] tracking self/parent flag
Summary:
Now that hackc is exporting the experimental flag directly in textual, we can parse it and use it in infer.
This diff improves TextualSil by checking the presence of the experimental flag for self and parent in trait at the module level, and thread it to all hack types.

Reviewed By: davidpichardie

Differential Revision: D46485726

fbshipit-source-id: 7709c748fc1ae25aed4b191dfe5cbecd5676a0d9