[hack] Rename hack-models cmdline flag to hack-builtin-models
Summary:
This is a preparatory step to introduce mode ways to specify additional
models.

The idea is that these models we usually bundle with infer and we don't
really want to override them in a production setting (maybe just in
tests). But we may want to add more model files on top (some in
textual, some perhaps in Hack).

Reviewed By: geralt-encore

Differential Revision: D46629386

fbshipit-source-id: 56e75ab6eb0d8b7cb88e3a4d33fc13d7867a05f2