[hack] Add an cmdline option --hack-models to supply extra model files
Summary:
The new option will accept a list of strings that will be treated as a list of .sil and .hack files
relative to the workspace root. These files will then be used to supply extra models.

The diff only adds a cmdline option. The actual wiring will happen in a follow up diff.

Reviewed By: geralt-encore

Differential Revision: D46629388

fbshipit-source-id: c81f783b0d0c02c008f05bbe7ef6eaac5e0bce5c